% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_charcoal.R
\name{interactive_charcoal}
\alias{interactive_charcoal}
\title{interactive charcoal filtering}
\usage{
interactive_charcoal(
  image,
  range_max_radius = 5,
  range_max_sigma = 5,
  resolution = 0.1,
  return_param = FALSE,
  scale
)
}
\arguments{
\item{image}{a magick image object}

\item{range_max_radius}{define maximum in slider of radius. must be positive.}

\item{range_max_sigma}{define maximum in slider of sigma. must be positive.}

\item{resolution}{resolution of slider}

\item{return_param}{If return_param is TRUE, returns values of radius and sigma. If return_param is FALSE, returns a magick image object.}

\item{scale}{geometry to be passed to image_scale function of magick package. image is scaled just for preview and result image is not scaled if scale is given.}
}
\value{
a magick image object or values of radius and sigma
}
\description{
Using image_charcoal of 'magick' interactively.
radius and sigma are parameters of image_charcoal. See reference manual of 'magick' for detail.
}
\examples{
\donttest{
interactive_charcoal(wizard)
}
}
\author{
Shota Ochi
}
