% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_crop.R
\name{interactive_crop}
\alias{interactive_crop}
\title{interactive cropping}
\usage{
interactive_crop(image, color = "white", return_param = FALSE)
}
\arguments{
\item{image}{a magick image object}

\item{color}{color of background. a valid color string such as "navyblue" or "#000080". "none" is not allowed.}

\item{return_param}{If return_param is TRUE, returns a value of geometry. If return_param is FALSE, returns a magick image object.}
}
\value{
a magick image object or a value of geometry.
}
\description{
Using image_crop of 'magick' interactively.
}
\examples{
\donttest{
interactive_crop(wizard)
}
}
\author{
Shota Ochi
}
