% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_despeckle.R
\name{interactive_despeckle}
\alias{interactive_despeckle}
\title{interactive despeckling}
\usage{
interactive_despeckle(
  image,
  range_max = 50,
  resolution = 1,
  return_param = FALSE,
  scale
)
}
\arguments{
\item{image}{a magick image object}

\item{range_max}{define maximum in slider. must be positive.}

\item{resolution}{resolution of slider}

\item{return_param}{If return_param is TRUE, returns value of times. If return_param is FALSE, returns a magick image object.}

\item{scale}{geometry to be passed to image_scale function of magick package. image is scaled just for preview and result image is not scaled if scale is given.}
}
\value{
a magick image object or value of times
}
\description{
Using image_despeckle of 'magick' interactively.
times is a parameter of image_despeckle. See reference manual of 'magick' for detail.
}
\examples{
\donttest{
if (interactive())
{
  interactive_despeckle(wizard)
}
}
}
\author{
Shota Ochi
}
