% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_emboss.R
\name{interactive_emboss}
\alias{interactive_emboss}
\title{interactive embossing}
\usage{
interactive_emboss(image, range_max_radius = 5, range_max_sigma = 5,
  resolution = 0.1, return_param = FALSE)
}
\arguments{
\item{image}{a magick image object}

\item{range_max_radius}{define maximum in slider of radius. must be positive.}

\item{range_max_sigma}{define maximum in slider of sigma. must be positive.}

\item{resolution}{resolution of slider}

\item{return_param}{If return_param is TRUE, returns values of radius and sigma. If return_param is FALSE, returns a magick image object.}
}
\value{
a magick image object or values of radius and sigma
}
\description{
Using image_emboss of 'magick' interactively.
radius and sigma are parameters of image_emboss. See reference manual of 'magick' for detail.
}
\examples{
\donttest{
interactive_emboss(wizard)
}
}
\author{
Shota Ochi
}
