% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_modulate.R
\name{interactive_modulate}
\alias{interactive_modulate}
\title{interactive modulating}
\usage{
interactive_modulate(image, range_max_brightness = 200,
  range_max_saturation = 200, range_max_hue = 200, resolution = 0.1,
  return_param = FALSE)
}
\arguments{
\item{image}{a magick image object}

\item{range_max_brightness}{define maximum in slider of brightness. must be positive.}

\item{range_max_saturation}{define maximum in slider of saturation. must be positive.}

\item{range_max_hue}{define maximum in slider of hue. must be positive.}

\item{resolution}{resolution of slider}

\item{return_param}{if TRUE, returns values of brightness and saturation and hue. if FALSE, returns magick image object.}
}
\value{
magick a image object or values of brightness and saturation
}
\description{
Using image_modulate of 'magick' interactively.
brightness and saturation and hue are parameters of image_modulate. see reference manual of 'magick' for detail.
}
\examples{
\donttest{
interactive_modulate(wizard)
}
}
\author{
Shota Ochi
}
