% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_motion_blur.R
\name{interactive_motion_blur}
\alias{interactive_motion_blur}
\title{interactive motion blurring}
\usage{
interactive_motion_blur(
  image,
  range_max_radius = 200,
  range_max_sigma = 200,
  range_max_angle = 360,
  resolution = 0.1,
  return_param = FALSE,
  scale
)
}
\arguments{
\item{image}{a magick image object}

\item{range_max_radius}{define maximum in slider of radius. must be positive.}

\item{range_max_sigma}{define maximum in slider of sigma. must be positive.}

\item{range_max_angle}{define maximum in slider of angle. must be positive.}

\item{resolution}{resolution of slider}

\item{return_param}{If return_param is TRUE, returns values of radius and sigma and angle. If return_param is FALSE, returns a magick image object.}

\item{scale}{geometry to be passed to image_scale function of magick package. image is scaled just for preview and result image is not scaled if scale is given.}
}
\value{
a magick image object or values of radius, sigma, and angle
}
\description{
Using image_motion_blur of 'magick' interactively.
radius and sigma and angle are parameters of image_motion_blur. See reference manual of 'magick' for detail.
}
\examples{
\donttest{
interactive_motion_blur(wizard)
}
}
\author{
Shota Ochi
}
