% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_reducenoise.R
\name{interactive_reducenoise}
\alias{interactive_reducenoise}
\title{interactive denoising}
\usage{
interactive_reducenoise(image, range_max = 30, resolution = 1,
  return_param = FALSE)
}
\arguments{
\item{image}{a magick image object}

\item{range_max}{define maximum in slider. must be positive.}

\item{resolution}{resolution of slider}

\item{return_param}{if TRUE, returns value of radius. if FALSE, returns magick image object.}
}
\value{
magick a image object or value of radius
}
\description{
Using image_reducenoise of 'magick' interactively.
radius is a parameter of image_reducenoise. see reference manual of 'magick' for detail.
}
\examples{
\dontrun{
interactive_reducenoise(wizard)
}
}
\author{
Shota Ochi
}
