% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_composite.R
\name{interactive_composite}
\alias{interactive_composite}
\title{interactive image compositing}
\usage{
interactive_composite(
  image,
  composite_image,
  operator = "atop",
  compose_args = "",
  resolution = 1,
  return_param = FALSE,
  scale
)
}
\arguments{
\item{image}{a magick image object}

\item{composite_image}{composition image}

\item{operator}{string with a composite operator}

\item{compose_args}{additional arguments needed for some composite operations}

\item{resolution}{resolution of slider}

\item{return_param}{If return_param is TRUE, returns values of offset. If return_param is FALSE, returns a magick image object.}

\item{scale}{geometry to be passed to image_scale function of magick package. image is scaled just for preview and result image is not scaled if scale is given.}
}
\value{
magick a image object or values of offset
}
\description{
Using image_composite of 'magick' interactively.
offset is a parameter of image_composite. see reference manual of 'magick' for detail.
}
\examples{
\donttest{
interactive_composite(wizard, rose)
}
}
\author{
Shota Ochi
}
