% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_despeckle.R
\name{interactive_despeckle}
\alias{interactive_despeckle}
\title{interactive despeckling}
\usage{
interactive_despeckle(image, range_max = 50, resolution = 1,
  return_param = FALSE)
}
\arguments{
\item{image}{a magick image object}

\item{range_max}{define maximum in slider. must be positive.}

\item{resolution}{resolution of slider}

\item{return_param}{if TRUE, returns value of times. if FALSE, returns magick image object.}
}
\value{
magick a image object or value of times
}
\description{
Using image_despeckle of 'magick' interactively.
times is a parameter of image_despeckle. see reference manual of 'magick' for detail.
}
\examples{
\donttest{
interactive_despeckle(wizard)
}
}
\author{
Shota Ochi
}
