% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_implode.R
\name{interactive_implode}
\alias{interactive_implode}
\title{interactive imploding}
\usage{
interactive_implode(
  image,
  range_max = 1,
  resolution = 0.1,
  return_param = FALSE,
  scale
)
}
\arguments{
\item{image}{a magick image object}

\item{range_max}{define maximum in slider. must be positive.}

\item{resolution}{resolution of slider}

\item{return_param}{If return_param is TRUE, returns value of factor. If return_param is FALSE, returns a magick image object.}

\item{scale}{geometry to be passed to image_scale function of magick package. image is scaled just for preview and result image is not scaled if scale is given.}
}
\value{
a magick image object or value of factor
}
\description{
Using image_implode of 'magick' interactively.
factor is a parameter of image_implode. See reference manual of 'magick' for detail.
}
\examples{
\donttest{
interactive_implode(wizard)
}
}
\author{
Shota Ochi
}
