% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_implode.R
\name{interactive_implode}
\alias{interactive_implode}
\title{interactive imploding}
\usage{
interactive_implode(image, range_max = 1, resolution = 0.1,
  return_param = FALSE)
}
\arguments{
\item{image}{a magick image object}

\item{range_max}{define maximum in slider. must be positive.}

\item{resolution}{resolution of slider}

\item{return_param}{if TRUE, returns value of factor. if FALSE, returns magick image object.}
}
\value{
a magick image object or value of factor
}
\description{
Using image_implode of 'magick' interactively.
factor is a parameter of image_implode. See reference manual of 'magick' for detail.
}
\examples{
\donttest{
interactive_implode(wizard)
}
}
\author{
Shota Ochi
}
