% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_oilpaint.R
\name{interactive_oilpaint}
\alias{interactive_oilpaint}
\title{interactive oil painting}
\usage{
interactive_oilpaint(image, range_max = 10, resolution = 0.1,
  return_param = FALSE)
}
\arguments{
\item{image}{a magick image object}

\item{range_max}{define maximum in slider. must be positive.}

\item{resolution}{resolution of slider}

\item{return_param}{if TRUE, returns value of radius. if FALSE, returns magick image object.}
}
\value{
magick a image object or value of radius
}
\description{
Using image_oilpaint of 'magick' interactively.
radius is a parameter of image_oilpaint. see reference manual of 'magick' for detail.
}
\examples{
\dontrun{
interactive_oilpaint(wizard)
}
}
\author{
Shota Ochi
}
