% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_threshold.R
\name{interactive_threshold}
\alias{interactive_threshold}
\title{interactive thresholding}
\usage{
interactive_threshold(image, type = c("black", "white"),
  channel = NULL, scale = 1, resolution = 0.1, return_thr = FALSE)
}
\arguments{
\item{image}{a magick image object}

\item{type}{type of thresholding, either one of lat, black or white}

\item{channel}{a value specifying which channel(s) to set}

\item{scale}{scale factor for size of image shown in plot. scale does not affect result.}

\item{resolution}{specifying resolution of slider. must be positive.}

\item{return_thr}{if TRUE, returns threshold value. if FALSE, returns magick image object.}
}
\value{
magick a image object or threshold value
}
\description{
Using image_threshold of 'magick' interactively.
type and channel are parameters of image_threshold. see reference manual of 'magick' for detail.
}
\examples{
\dontrun{
interactive_threshold(wizard)
}
}
\author{
Shota Ochi
}
