% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_threshold.R
\name{interactive_threshold}
\alias{interactive_threshold}
\title{interactive thresholding}
\usage{
interactive_threshold(
  image,
  type = c("black", "white"),
  channel = NULL,
  resolution = 0.1,
  return_param = FALSE,
  scale
)
}
\arguments{
\item{image}{a magick image object}

\item{type}{type of thresholding, either one of lat, black or white}

\item{channel}{a value specifying which channel(s) to set}

\item{resolution}{resolution of slider}

\item{return_param}{If return_param is TRUE, returns threshold value. If return_param is FALSE, returns a magick image object.}

\item{scale}{geometry to be passed to image_scale function of magick package. image is scaled just for preview and result image is not scaled if scale is given.}
}
\value{
a magick image object or threshold value
}
\description{
Using image_threshold of 'magick' interactively.
threshold is a parameter of image_threshold. See reference manual of 'magick' for detail.
}
\examples{
\donttest{
if (interactive())
{
  interactive_threshold(wizard)
}
}
}
\author{
Shota Ochi
}
