% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{package}
\name{magickGUI}
\alias{magickGUI}
\alias{magickGUI-package}
\title{magickGUI: GUI tools for interactive image processing with 'magick'}
\description{
magickGUI enables us to use the functions of the package 'magick' interactively.
}
