% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{package}
\name{magickGUI}
\alias{magickGUI}
\alias{magickGUI-package}
\title{magickGUI: GUI tools for interactive image processing with 'magick'}
\description{
magickGUI enables us to use the functions of the package 'magick' interactively.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ShotaOchi/magickGUI}
  \item Report bugs at \url{https://github.com/ShotaOchi/magickGUI/issues}
}

}
\author{
\strong{Maintainer}: Shota Ochi \email{shotaochi1990@gmail.com}

}
